requirements.txt
streamlit>=1.38.0
google-genai>=1.0.0
python-dotenv>=1.0.1
pandas>=2.2.0
